requirements.txt
face_recognition_models
Click>=6.0
dlib>=19.3.0
Flask==1.1.1
numpy
Pillow
scipy>=0.17.0
